+++ addon/components/olv-toolbar.js
@@ -133,6 +133,16 @@ export default FlexberryBaseComponent.extend({
     @default null
   */
   filterText: null,
+
+  /**
+    Indicates that the `flexberry-objectlistview` component is used for the `flexberry-lookup` component.
+
+    @property inLookup
+    @type Boolean
+    @default false
+  */
+  inLookup: false,
+
   /**
     Used to link to objectListView with same componentName.
 
@@ -141,6 +151,14 @@ export default FlexberryBaseComponent.extend({
     @default ''
   */
   componentName: '',
+
+  /**
+    The name of the `flexberry-lookup` component for which the `flexberry-objectlistview` component is used.
+
+    @property lookupComponentName
+    @type String
+  */
+  lookupComponentName: undefined,
 
   /**
     The flag to specify whether the delete button is enabled.
@@ -461,6 +479,18 @@ export default FlexberryBaseComponent.extend({
   _infoModalDialog: null,
 
   /**
+    The name of the component for themodal window.
+    If the `flexberry-objectlistview` component is used for the `flexberry-lookup` component, we must pass the name of the `flexberry-lookup` component to the modal window.
+
+    @private
+    @property _componentNameForModalWindow
+    @type String
+  */
+  _componentNameForModalWindow: Ember.computed('inLookup', 'componentName', 'lookupComponentName', function () {
+    return this.get('inLookup') ? this.get('lookupComponentName') : this.get('componentName');
+  }),
+
+  /**
    Shows info modal dialog.
 
    @method showInfoModalDialog
@@ -621,7 +651,7 @@ export default FlexberryBaseComponent.extend({
     */
     showConfigDialog(settingName) {
       assert('showConfigDialog:: componentName is not defined in flexberry-objectlistview component', this.componentName);
-      this.get('modelController').send('showConfigDialog', this.componentName, settingName);
+      this.get('modelController').send('showConfigDialog', this.get('_componentNameForModalWindow'), settingName);
     },
 
     /**
@@ -632,7 +662,7 @@ export default FlexberryBaseComponent.extend({
     */
     showAdvLimitDialog(settingName) {
       assert('showAdvLimitDialog:: componentName is not defined in flexberry-objectlistview component', this.componentName);
-      this.get('modelController').send('showAdvLimitDialog', this.componentName, settingName);
+      this.get('modelController').send('showAdvLimitDialog', this.get('_componentNameForModalWindow'), settingName);
     },
 
     /**
@@ -663,7 +693,7 @@ export default FlexberryBaseComponent.extend({
     showExportDialog(settingName, immediateExport) {
       let settName = settingName ? 'ExportExcel/' + settingName : settingName;
       assert('showExportDialog:: componentName is not defined in flexberry-objectlistview component', this.componentName);
-      this.get('modelController').send('showConfigDialog', this.componentName, settName, true, immediateExport);
+      this.get('modelController').send('showConfigDialog', this.get('_componentNameForModalWindow'), settName, true, immediateExport);
     },
 
     /**